Move tests.
